+++ src/main/java/com/isroot/stash/plugin/YaccHook.java
@@ -5,11 +5,13 @@ import com.atlassian.stash.hook.repository.PreReceiveRepositoryHook;
 import com.atlassian.stash.hook.repository.RepositoryHookContext;
 import com.atlassian.stash.repository.RefChange;
 import com.atlassian.stash.repository.RefChangeType;
+import com.atlassian.stash.setting.Settings;
 import com.google.common.collect.Lists;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
+import java.io.PrintWriter;
 import java.util.Collection;
 import java.util.List;
 
@@ -46,6 +48,7 @@ public final class YaccHook implements PreReceiveRepositoryHook
                              @Nonnull Collection<RefChange> refChanges, @Nonnull HookResponse hookResponse)
     {
         List<String> errors = Lists.newArrayList();
+        Settings settings = repositoryHookContext.getSettings();
 
         for (RefChange rf : refChanges)
         {
@@ -54,7 +57,8 @@ public final class YaccHook implements PreReceiveRepositoryHook
                 continue;
             }
 
-            errors.addAll(yaccService.checkRefChange(repositoryHookContext.getRepository(), repositoryHookContext.getSettings(), rf));
+            errors.addAll(yaccService.checkRefChange(repositoryHookContext.getRepository(),
+                    settings, rf));
         }
 
         if (errors.isEmpty())
@@ -65,7 +69,7 @@ public final class YaccHook implements PreReceiveRepositoryHook
         }
         else
         {
-            hookResponse.err().println(ERROR_BEARS);
+            printHeader(settings, hookResponse.err());
 
             for (String error : errors)
             {
@@ -76,9 +80,35 @@ public final class YaccHook implements PreReceiveRepositoryHook
 
             hookResponse.err().println();
 
+            printFooter(settings, hookResponse.err());
+
             log.debug("push rejected");
 
             return false;
+        }
+    }
+
+    private void printHeader(Settings settings, PrintWriter writer)
+    {
+        String header = settings.getString("errorMessageHeader");
+
+        if(header == null || header.isEmpty())
+        {
+            // sford: long live the error bears
+            header = ERROR_BEARS;
+        }
+
+        writer.println(header);
+        writer.println();
+    }
+
+    private void printFooter(Settings settings, PrintWriter writer)
+    {
+        String footer = settings.getString("errorMessageFooter");
+        if(footer != null && !footer.isEmpty())
+        {
+            writer.println(footer);
+            writer.println();
         }
     }
 }