+++ src/test/java/ut/com/isroot/stash/plugin/YaccHookTest.java
@@ -15,11 +15,11 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import java.io.PrintWriter;
+import java.io.StringWriter;
 
 import static org.fest.assertions.api.Assertions.assertThat;
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
@@ -32,6 +32,9 @@ public class YaccHookTest
 	@Mock private YaccService yaccService;
 	@Mock private HookResponse hookResponse;
 	@Mock private RepositoryHookContext repositoryHookContext;
+    @Mock private Settings settings;
+
+    private StringWriter errorMessage;
 
 	private YaccHook yaccHook;
 
@@ -42,7 +45,10 @@ public class YaccHookTest
 
 		yaccHook = new YaccHook(yaccService);
 
-		when(hookResponse.err()).thenReturn(mock(PrintWriter.class));
+        errorMessage = new StringWriter();
+
+		when(hookResponse.err()).thenReturn(new PrintWriter(errorMessage));
+        when(repositoryHookContext.getSettings()).thenReturn(settings);
 	}
 
 	@Test
@@ -51,7 +57,7 @@ public class YaccHookTest
 		RefChange refChange = mock(RefChange.class);
 		when(refChange.getType()).thenReturn(RefChangeType.DELETE);
 
-	 	boolean allowed = yaccHook.onReceive(null, Lists.newArrayList(refChange), null);
+	 	boolean allowed = yaccHook.onReceive(repositoryHookContext, Lists.newArrayList(refChange), null);
 		assertThat(allowed).isTrue();
 		verifyZeroInteractions(yaccService);
 	}
@@ -74,10 +80,46 @@ public class YaccHookTest
 
 		yaccHook.onReceive(repositoryHookContext, Lists.newArrayList(mock(RefChange.class)), hookResponse);
 
-		verify(hookResponse.err()).println(YaccHook.ERROR_BEARS);
-		verify(hookResponse.err()).println("error1");
-		verify(hookResponse.err()).println("error2");
+        assertThat(errorMessage.toString())
+                .isEqualTo(YaccHook.ERROR_BEARS + "\n\nerror1\nerror2\n\n");
 	}
 
+    @Test
+    public void testOnReceive_defaultHeaderDisplayedIfErrorMessageHeaderIsEmpty()
+    {
+        when(yaccService.checkRefChange(any(Repository.class), any(Settings.class), any(RefChange.class)))
+                .thenReturn(Lists.newArrayList("error1"));
 
+        when(settings.getString("errorMessageHeader")).thenReturn("");
+
+        yaccHook.onReceive(repositoryHookContext, Lists.newArrayList(mock(RefChange.class)), hookResponse);
+
+        assertThat(errorMessage.toString()).startsWith(YaccHook.ERROR_BEARS);
+    }
+
+    @Test
+    public void testOnReceive_nonEmptyErrorMessageHeaderReplacesDefaultHeader()
+    {
+        when(yaccService.checkRefChange(any(Repository.class), any(Settings.class), any(RefChange.class)))
+                .thenReturn(Lists.newArrayList("error1"));
+
+        when(settings.getString("errorMessageHeader")).thenReturn("Custom Header");
+
+        yaccHook.onReceive(repositoryHookContext, Lists.newArrayList(mock(RefChange.class)), hookResponse);
+
+        assertThat(errorMessage.toString()).isEqualTo("Custom Header\n\nerror1\n");
+    }
+
+    @Test
+    public void testOnReceive_errorMessageFooterAddedToEndOfOutput()
+    {
+        when(yaccService.checkRefChange(any(Repository.class), any(Settings.class), any(RefChange.class)))
+                .thenReturn(Lists.newArrayList("error1"));
+
+        when(settings.getString("errorMessageFooter")).thenReturn("Custom Footer");
+
+        yaccHook.onReceive(repositoryHookContext, Lists.newArrayList(mock(RefChange.class)), hookResponse);
+
+        assertThat(errorMessage.toString()).endsWith("\nCustom Footer\n\n");
+    }
 }